+++ src/steps/constants.ts
@@ -6,16 +6,41 @@ import {
   StepRelationshipMetadata,
 } from '@jupiterone/integration-sdk-core';
 
+export const ACCOUNT_ENTITY_KEY = 'entity:account';
+
 export const Steps = {
-  PROGRAM: 'build-program',
+  ACCOUNT: 'fetch-account',
+  ORGANIZATION: 'fetch-organization',
+  PROGRAMS: 'build-programs',
+  SERVICE: 'fetch-service',
+  ASSESSMENT: 'fetch-assessments',
+  PROGRAM_ASSETS: 'fetch-program-assets',
   REPORTS: 'fetch-reports',
-  PROGRAM_REPORTS_RELATIONSHIPS: 'build-program-reports-relationships',
+  PROGRAM_ASSETS_REPORTS_RELATIONSHIPS:
+    'build-program-assets-reports-relationships',
 };
 
 export const Entities: Record<
-  'PROGRAM' | 'REPORT' | 'CVE' | 'CWE',
+  | 'ACCOUNT'
+  | 'PROGRAM'
+  | 'PROGRAM_ASSET'
+  | 'ORGANIZATION'
+  | 'ASSESSMENT'
+  | 'REPORT'
+  | 'CVE'
+  | 'CWE',
   StepEntityMetadata
 > = {
+  ACCOUNT: {
+    resourceName: 'Account',
+    _type: 'hackerone_account',
+    _class: ['Account'],
+  },
+  ORGANIZATION: {
+    resourceName: 'Organization',
+    _type: 'hackerone_organization',
+    _class: ['Organization'],
+  },
   PROGRAM: {
     resourceName: 'Service',
     _type: 'hackerone_program',
@@ -28,6 +53,16 @@ export const Entities: Record<
       },
       required: ['category', 'handle'],
     },
+  },
+  PROGRAM_ASSET: {
+    resourceName: 'Program Asset',
+    _type: 'hackerone_program_asset',
+    _class: ['Entity'], // TBD: A better fitting class
+  },
+  ASSESSMENT: {
+    resourceName: 'Assessment',
+    _type: 'hackerone_assessment',
+    _class: ['Assessment'],
   },
   REPORT: {
     resourceName: 'Finding',
@@ -57,12 +92,61 @@ export const Entities: Record<
 };
 
 export const Relationships: Record<
-  'PROGRAM_REPORTED_FINDING',
+  | 'PROGRAM_REPORTED_FINDING'
+  | 'PROGRAM_PERFORMED_ASSESSMENT'
+  | 'PROGRAM_SCANS_PROGRAM_ASSET'
+  | 'ACCOUNT_HAS_ORGANIZATION'
+  | 'ACCOUNT_HAS_PROGRAM_ASSET'
+  | 'ACCOUNT_HAS_PROGRAM'
+  | 'PROGRAM_ASSET_HAS_FINDING'
+  | 'ORGANIZATION_HAS_PROGRAM',
   StepRelationshipMetadata
 > = {
   PROGRAM_REPORTED_FINDING: {
     _type: 'hackerone_program_reported_finding',
     sourceType: Entities.PROGRAM._type,
+    _class: RelationshipClass.IDENTIFIED,
+    targetType: Entities.REPORT._type,
+  },
+  ACCOUNT_HAS_ORGANIZATION: {
+    _type: 'hackerone_account_has_organization',
+    sourceType: Entities.ACCOUNT._type,
+    _class: RelationshipClass.HAS,
+    targetType: Entities.ORGANIZATION._type,
+  },
+  ACCOUNT_HAS_PROGRAM_ASSET: {
+    _type: 'hackerone_account_has_program_asset',
+    sourceType: Entities.ACCOUNT._type,
+    _class: RelationshipClass.HAS,
+    targetType: Entities.PROGRAM_ASSET._type,
+  },
+  ACCOUNT_HAS_PROGRAM: {
+    _type: 'hackerone_account_has_program',
+    sourceType: Entities.ACCOUNT._type,
+    _class: RelationshipClass.HAS,
+    targetType: Entities.PROGRAM._type,
+  },
+  ORGANIZATION_HAS_PROGRAM: {
+    _type: 'hackerone_organization_has_program',
+    sourceType: Entities.ORGANIZATION._type,
+    _class: RelationshipClass.HAS,
+    targetType: Entities.PROGRAM._type,
+  },
+  PROGRAM_PERFORMED_ASSESSMENT: {
+    _type: 'hackerone_program_performed_assessment',
+    sourceType: Entities.PROGRAM._type,
+    _class: RelationshipClass.PERFORMED,
+    targetType: Entities.ASSESSMENT._type,
+  },
+  PROGRAM_SCANS_PROGRAM_ASSET: {
+    _type: 'hackerone_program_scans_asset',
+    sourceType: Entities.PROGRAM._type,
+    _class: RelationshipClass.SCANS,
+    targetType: Entities.PROGRAM_ASSET._type,
+  },
+  PROGRAM_ASSET_HAS_FINDING: {
+    _type: 'hackerone_program_asset_has_report',
+    sourceType: Entities.PROGRAM_ASSET._type,
     _class: RelationshipClass.HAS,
     targetType: Entities.REPORT._type,
   },
@@ -73,14 +157,14 @@ export const MappedRelationships: Record<
   StepMappedRelationshipMetadata
 > = {
   FINDING_EXPLOITS_WEAKNESS: {
-    _type: 'hackerone_finding_exploits_weakness',
+    _type: 'hackerone_report_exploits_weakness',
     sourceType: Entities.REPORT._type,
     _class: RelationshipClass.HAS,
     targetType: Entities.CWE._type,
     direction: RelationshipDirection.FORWARD,
   },
   FINDING_IS_VULNERABILITY: {
-    _type: 'hackerone_finding_is_vulnerability',
+    _type: 'hackerone_report_is_vulnerability',
     sourceType: Entities.REPORT._type,
     _class: RelationshipClass.HAS,
     targetType: Entities.CVE._type,